+++ NEWS.md
@@ -8,6 +8,7 @@
 - Remove unused GRANTING_INSTITUTION, DISSERTATION_NOTE and DISSERTATION_YEAR properties. Create DissertationLabelGenerator[MODLD-1014](https://folio-org.atlassian.net/browse/MODLD-1014)
 - Correct DIMENSIONS property URI [MODLD-1024](https://folio-org.atlassian.net/browse/MODLD-1024)
 - Get rid of RESOURCE_PREFERRED property [MODLD-1028](https://folio-org.atlassian.net/browse/MODLD-1028)
+- Create ProviderEventLabelGenerator [MODLD-1020](https://folio-org.atlassian.net/browse/MODLD-1020)
 
 ## v1.9.9 (13-02-2026)
 - Update Mapping for 300$b & $e (Physical description & Accompanying material) [MODLD-508](https://folio-org.atlassian.net/browse/MODLD-508)